Move sentry to a controller
