+++ unity-renderer/Assets/DCLPlugins/SentryPlugin/SentryPlugin.cs
@@ -1,58 +1,17 @@
 using DCL;
+using DCLPlugins.SentryPlugin;
 using Sentry;
 using UnityEngine;
 public class SentryPlugin : IPlugin
 {
+    private readonly SentryController controller;
     public SentryPlugin()
     {
-        DataStore.i.player.playerGridPosition.OnChange += PlayerGridPositionOnOnChange;
-        DataStore.i.player.otherPlayers.OnAdded += OtherPlayersOnChanged;
-        DataStore.i.player.otherPlayers.OnRemoved += OtherPlayersOnChanged;
-        DataStore.i.player.lastTeleportPosition.OnChange += LastTeleportPositionOnOnChange;
-        DataStore.i.realm.realmName.OnChange += RealmNameOnOnChange;
-    }
-
-    private void RealmNameOnOnChange(string current, string previous)
-    {
-        SentrySdk.ConfigureScope(scope =>
-        {
-            scope.SetTag("Current Realm", current);
-            scope.SetTag("Previous realm", previous);
-        });
-    }
-
-    private void LastTeleportPositionOnOnChange(Vector3 current, Vector3 previous)
-    {
-        SentrySdk.ConfigureScope(scope =>
-        {
-            scope.SetExtra("Current Teleport Position", $"{current.x},{current.y}");
-            scope.SetExtra("Last Teleport Position", $"{previous.x},{previous.y}");
-        });
-    }
-
-    private void OtherPlayersOnChanged(string _, Player __)
-    {
-        SentrySdk.ConfigureScope(scope =>
-        {
-            scope.SetExtra("Total Other Players", $"{DataStore.i.player.otherPlayers.Count()}");
-        });
-    }
-
-    private void PlayerGridPositionOnOnChange(Vector2Int current, Vector2Int previous)
-    {
-        SentrySdk.ConfigureScope(scope =>
-        {
-            scope.SetTag("Current Position", $"{current.x},{current.y}");
-            scope.SetTag("Previous Position", $"{previous.x},{previous.y}");
-        });
+        controller = new SentryController(DataStore.i);
     }
 
     public void Dispose()
     {
-        DataStore.i.player.playerGridPosition.OnChange -= PlayerGridPositionOnOnChange;
-        DataStore.i.player.otherPlayers.OnAdded -= OtherPlayersOnChanged;
-        DataStore.i.player.otherPlayers.OnRemoved -= OtherPlayersOnChanged;
-        DataStore.i.player.lastTeleportPosition.OnChange -= LastTeleportPositionOnOnChange;
-        DataStore.i.realm.realmName.OnChange -= RealmNameOnOnChange;
+        controller?.Dispose();
     }
 }
